+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 ### Added
 
+- Forcefield templates:
+  - `selected_pair_forcefield`
 - Potentials:
   - `polybell_potential`
 - Misc: